users/views/delete.py
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework.permissions import IsAuthenticated
from rest_framework import status

class DeleteView(APIView):
    permission_classes = [IsAuthenticated]

    def delete(self, request):
        user = request.user
        user.delete()
        return Response({
            'detail': 'Usuário deletado com sucesso!'
        }, status=status.HTTP_204_NO_CONTENT)