pyproject.toml
[project]
name = "newback"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AndreRicardoSc", email = "andrenarutto7@gmail.com"},
]
dependencies = ["django>=4.2.24", "djangorestframework>=3.16.1", "djangorestframework-simplejwt>=5.5.1", "Pillow>=11.3.0", "drf-spectacular>=0.28.0", "django-cors-headers>=4.9.0", "psycopg2-binary>=2.9.10", "dj-database-url>=3.0.1"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
